+++ agent/testsrc/com/amazon/ebook/booklet/reader/impl/whisperstore/WhisperStoreKwisUtils.java
@@ -1,0 +1,9 @@
+package com.amazon.ebook.booklet.reader.impl.whisperstore;
+
+public final class WhisperStoreKwisUtils {
+    private WhisperStoreKwisUtils() {}
+
+    public static boolean Ls() {
+        return false;
+    }
+}