+++ agent/testsrc/com/amazon/ebook/booklet/reader/sdk/content/annotation/Annotation.java
@@ -1,0 +1,12 @@
+package com.amazon.ebook.booklet.reader.sdk.content.annotation;
+
+public interface Annotation {
+    Integer jm();
+    Object jh();
+    Object jd();
+    Object Cf();
+    Object L();
+    Object Cd();
+    AnnotationExport Ci();
+    JournalType CS();
+}